==> notebook/01_explore_Ragni_Verser.ipynb <==
import pandas as pd

df1 = pd.read_csv("../data/ressources/Ragni2016.csv")
df2 = pd.read_csv("../data/ressources/Veser2018.csv")

## Analyse de la base Ragni 2016

df1.head()

La base de données comporte les données suivantes:
- id: l'id du participant
- sequence: un id pour chaque type de syllogisme donnée à chaque participant
- task: les deux premisses du syllogisme proposé
- choices: les 8 réponses proposées (dont NVC : no valid conclusion)
- response: la réponse du participant
- gender, age : informations sur le participant
- response_type, domain: varibles contantes

df1.shape
# Le jeu contient 8896 lignes, c-a-d 8896 syllogismes ont été soumis

len(pd.unique(df1["id"]))
# Il y a 139 participants différents, chacun on été soumis à 64 syllogism, 1 par forme.

len(pd.unique(df1["task"]))
# Certains syllogismes ont été proposés plusieurs fois, en tout 6327 syllogisms différents ont été soumis.

print(df1["task"].value_counts().max())
# Au maximum, un même syllogisme a été proposé 6 fois

## Analyse de la base Veser 2018

df2.head()

Le jeu comporte les mêmes colonnes

df2.shape
# Le jeu contient 2058 ligne, c-a-d 2058 syllogismes ont été soumis

len(pd.unique(df2["id"]))
# Il y a 34 participants, chacun on été soumis à 64 syllogism, 1 par forme.

len(pd.unique(df2["task"]))
# Certains syllogismes ont été proposés plusieurs fois, en tout 383 syllogisms différents ont été soumis.

print(df2["task"].value_counts().max())
# Au maximum, un même syllogisme a été proposé 12 fois
