==> notebook/02_Preparation_données.ipynb <==
import pandas as pd
# package contenant la classe Syllogisme définie dans le cadre du projet 
# qui permet de manipuler les syllogismes et les formats des bdd.
import syllogism as sy 

from utils import table_de_verite, choice_form, implicatur



###----------- Import des bases

df1 = pd.read_csv("../data/ressources/Ragni2016.csv")
df2 = pd.read_csv("../data/ressources/Veser2018.csv")

###----------- Preparation des données en vue de l'analyse

#########----- Création de l'id unique regroupant id et sequence
df1 = df1.assign(id_seq=lambda df: df1.id.apply(str).str.cat(df1.sequence.apply(str),sep="_") + "_R")
df2 = df2.assign(id_seq=lambda df: df2.id.apply(str).str.cat(df2.sequence.apply(str),sep="_") + "_V")


#########----- Concaténation des deux tables

df_concat = pd.concat([df1,df2])

print(len(pd.unique(df_concat["id"])))
print(len(pd.unique(df_concat["task"])))
print(df_concat.shape)

df_concat = df_concat.drop(['id', 'sequence'], axis=1)


df_concat.to_csv("../data/intermediate/df_concat.csv",index=False)


#########----- Pour ne pas entrainer deux fois un même syllogisme, on supprimer les doublons
df_drop = df_concat.drop_duplicates(subset=['task','choices'])


#########----- Transformation des données en phrases valides:
df_drop['sentenced'] = df_drop.task.apply(lambda x : sy.Syllogism(x).sentenced)
df_drop['choice_str'] = df_drop[["task","choices"]].apply(lambda x: sy.Syllogism(x[0]).choice_to_str(x[1]), axis=1)

df_drop[['sentenced',"id_seq"]].to_csv("./article/data/idseq_sentence.csv" ,index=False)

#########----- Selection des colonnes
df_to_train = df_drop[["id_seq","sentenced","choice_str"]]

print(df_to_train.shape)
#########----- enregristement de la base préparée
#df_to_train.to_csv("./data/intermediate/df_to_train.csv" ,index=False)

df_to_train.tail(10)

print(len(pd.unique(df_to_train["id"])))
print(len(pd.unique(df_to_train["task"])))

# ###----------- Création de la base de données pour l'analyse des résultats
df_drop['task_form'] = df_drop.task.apply(lambda x : sy.Syllogism(x).full_form)
df_drop['human_response'] = df_drop[["task","response"]].apply(lambda x: sy.Syllogism(x[0]).evaluate_conclusion(x[1]), axis=1)
df_drop['has_conclusion'] = df_drop.task.apply(lambda x : sy.Syllogism(x).has_conclusion)
df_drop['human_response_index'] = df_drop.human_response.apply(lambda x : choice_form.index(x[0]))


def index_conclusion(task_form):
    
    index_ccl_list=[]
    # on parcourt les réponses possibles
    for i in choice_form:
        # on parcourt les bonnes réponses
        for ccl in table_de_verite[task_form]:
            if ccl == "NVC":
                return [8]
            elif i == ccl:
                index_ccl_list.append(choice_form.index(i))
    
    return index_ccl_list

df_drop['index_conclusion'] = df_drop.task_form.apply(lambda x : index_conclusion(x))


df_drop = df_drop[["id_seq", "task_form","human_response","has_conclusion","index_conclusion","choice_str",'human_response_index']]
df_drop.head()

df_drop['implicatur'] = df_drop.task_form.apply(lambda x : True if x in implicatur else False)

df_drop.to_csv("../data/intermediate/df_conclusion.csv" ,index=False)
