==> notebook/05_result_analysis.ipynb <==
## Set up

import pandas as pd
import ast
import syllogism as sy
from select_best import *


# On récupère les prédictions des trois modèles
df_bert = pd.read_csv("../data/intermediate/df_trained_bert.csv").rename(columns={"choice_pred": "choice_bert_pred"})
df_mnli = pd.read_csv("../data/intermediate/df_trained_MNLI.csv")
df_few_shot = pd.read_csv("../data/intermediate/df_trained_few_shot.csv")

# We drop useless column
df_bert = df_bert.drop(["choice_str","sentenced"], axis=1)
df_mnli = df_mnli.drop(["choice_str","sentenced"], axis=1)

## fusion des dataframes

# We import the dataframe with the orginal choices and syllogism form.
df_conclusion = pd.read_csv("../data/intermediate/df_conclusion.csv")

# merging of the dataframe
df_final = pd.merge(df_bert, df_mnli, on="id_seq", suffixes=('_prediction', '_chelma'),how="inner")
df_final = pd.merge(df_final, df_few_shot[["few_shot_pred","id_seq" ]], on=["id_seq"],how="left")
df_final = pd.merge(df_final, df_conclusion, on=["id_seq"])

df_final.to_csv("../data/intermediate/df_merge_results.csv" ,index=False)

## Calcul des taux de bonnes réponses de chaque modèle

models = {
    "bert_00" : {"model_type": "bert", "func": bert, "seuil":0.01},
    "bert_0999996" : {"model_type": "bert", "func": bert, "seuil":0.999996},
    "bert_0999997" : {"model_type": "bert", "func": bert, "seuil":0.999997},
    "bert_09999965" : {"model_type": "bert", "func": bert, "seuil":0.9999965},
    "bert_09999968" : {"model_type": "bert", "func": bert, "seuil":0.9999968},
    "bert_09999972" : {"model_type": "bert", "func": bert, "seuil":0.9999972},
    "bert_09999974" : {"model_type": "bert", "func": bert, "seuil":0.9999974},


    "mnli_00" : {"model_type": "mnli", "func": mnli, "seuil":0.001},
    "mnli_03" : {"model_type": "mnli", "func": mnli, "seuil":3},
    "mnli_05" : {"model_type": "mnli", "func": mnli, "seuil":5},
    "mnli_10" : {"model_type": "mnli", "func": mnli, "seuil":10},
    "mnli_20" : {"model_type": "mnli", "func": mnli, "seuil":20},
    "mnli_30" : {"model_type": "mnli", "func": mnli, "seuil":30},
    "mnli_40" : {"model_type": "mnli", "func": mnli, "seuil":40},
    "mnli_50" : {"model_type": "mnli", "func": mnli, "seuil":50},
    "mnli_60" : {"model_type": "mnli", "func": mnli, "seuil":60},
    "mnli_70" : {"model_type": "mnli", "func": mnli, "seuil":70},
    "mnli_80" : {"model_type": "mnli", "func": mnli, "seuil":80},
    "mnli_90" : {"model_type": "mnli", "func": mnli, "seuil":90},
    "mnli_3_options" : {"model_type": "mnli", "func": mnli_3_options, "seuil":False},
     
     "fewshot" : {"model_type": "few_shot", "func": few_shot, "seuil":False}
    
}


for key,value in models.items():
    if value["model_type"] == "few_shot":
        df_final[f'best_{key}_index'] = df_final[[f'{value["model_type"]}_pred', "choice_str"]].apply(lambda x: value["func"](x[0],x[1]), axis=1)
    elif value["seuil"]:
        df_final[f'best_{key}_index'] = df_final[f'choice_{value["model_type"]}_pred'].apply(lambda x: value["func"](x,value["seuil"]))
    else:
        df_final[f'best_{key}_index'] = df_final[f'choice_{value["model_type"]}_pred'].apply(lambda x: value["func"](x)) 

    # On détermine si les réponses sont valides ou non
    df_final[f'{key}success'] = df_final[[f'best_{key}_index',"index_conclusion"]].apply(lambda x: 1 if x[0] in ast.literal_eval(x[1]) else 0, axis=1)
    # On détermine la similarité entre la réponse humaine et celle du modèle
    df_final[f'similarity_{key}'] = df_final[["human_response_index",f'best_{key}_index']].apply(lambda x: 1 if x[0]==x[1] else 0, axis=1)

df_final["succes_human"]= df_final.human_response.apply(lambda x : 1 if ast.literal_eval(x)[1]==True else 0 )


df_final.to_csv("../data/intermediate/df_global_id.csv" ,index=False)

df_final.head()

# Deprecated

df_final['similarity_bert'] = df_final[["human_response_index","best_bert_index"]].apply(lambda x: 1 if x[0]==x[1] else 0, axis=1)
df_final['similarity_mnli'] = df_final[["human_response_index","best_mnli_index"]].apply(lambda x: 1 if x[0]==x[1] else 0, axis=1)
df_final['similarity_fewshot'] = df_final[["human_response_index","best_fewshot_index"]].apply(lambda x: 1 if x[0]==x[1] else 0, axis=1)

# On détermine si les réponses sont valides ou non
df_final['bert_succes'] = df_final[["best_bert_index","index_conclusion"]].apply(lambda x: 1 if x[0] in ast.literal_eval(x[1]) else 0, axis=1)
df_final['mnli_succes'] = df_final[["best_mnli_index","index_conclusion"]].apply(lambda x: 1 if x[0] in ast.literal_eval(x[1]) else 0, axis=1)
df_final['fewshot_succes'] = df_final[["best_fewshot_index","index_conclusion"]].apply(lambda x: 1 if x[0] in ast.literal_eval(x[1]) else 0, axis=1)
df_final["succes_human"]= df_final.human_response.apply(lambda x : 1 if ast.literal_eval(x)[1]==True else 0 )


df_final.to_csv("../data/intermediate/df_global_id.csv" ,index=False)

# On sépare les syllogisme qui ont une conclusion valide des autres
df_valid= df_to_analyse[df_to_analyse['has_conclusion'] == True]
df_unvalid = df_to_analyse[df_to_analyse['has_conclusion'] == False]

# On calcul le taux de succès de chaque modèle pour chaque
df_result_human = df_valid[['task_form', 'succes_human']].groupby(['task_form']).mean().merge(df_valid[['task_form', 'succes_human']].groupby(['task_form']).count(), on="task_form")
df_result_human = df_result_human.drop(["succes_human_y"], axis=1)

df_result_prediction = df_valid[['task_form', 'succes_prediction']].groupby(['task_form']).mean().merge(df_valid[['task_form', 'succes_prediction']].groupby(['task_form']).count(), on="task_form")
df_result_prediction = df_result_prediction.drop(["succes_prediction_y"], axis=1)

df_result_chelma = df_valid[['task_form', 'succes_chelma']].groupby(['task_form']).mean().merge(df_valid[['task_form', 'succes_chelma']].groupby(['task_form']).count(), on="task_form")
df_result_chelma = df_result_chelma.drop(["succes_chelma_y"], axis=1)

df_result_few_shot = df_valid[['task_form', 'succes_few_shot']].groupby(['task_form']).mean().merge(df_valid[['task_form', 'succes_few_shot']].groupby(['task_form']).count(), on="task_form")
df_result_few_shot = df_result_few_shot.drop(["succes_few_shot_y"], axis=1)

df_count = df_valid.groupby(["task_form"]).size().reset_index(name='counts')

# Enfin on fusionne les résultats des différents modèles dans une meme table
df_result = pd.merge(df_result_human, df_result_prediction, on=["task_form"])
df_result = pd.merge(df_result, df_result_chelma, on=["task_form"])
df_result = pd.merge(df_result, df_result_few_shot, on=["task_form"])
df_result['Syllogism'] = df_result.index
df_result = pd.merge(df_result, df_count, on=["task_form"])



df_result = df_result.rename(columns={"succes_human_x": "Human", "succes_prediction_x":"Bert_simple", "succes_chelma_x":"Bert_Chelma", "succes_few_shot_x":"Bart_MNLI" })
df_result = df_result[['Syllogism','Human','Bert_simple','Bert_Chelma','Bart_MNLI',"counts"]]

df_result

df_result.to_csv("../data/results/df_result_29_01_2022.csv" ,index=False)
