==> notebook/10_result_analysis.ipynb <==
import pandas as pd
import numpy as np

df_valid = pd.read_csv("../data/results/result_valid_30_01_22.csv").drop(['Unnamed: 0'], axis=1)
df_unvalid = pd.read_csv("../data/results/result_unvalid_30_01_22.csv").drop(['Unnamed: 0'], axis=1)


df_valid.columns

df_cor_valid = pd.DataFrame({'succes_mnli_max': [df_valid['succes_human'].corr(df_valid['succes_mnli_max'])]})
for col in df_valid.columns.drop(["succes_human","succes_mnli_max","task_form"]):
    df_cor_valid[col]= df_valid['succes_human'].corr(df_valid[col])
df_cor_valid.to_csv("../data/results/valid_cor_04_03_22.csv")
df_cor_valid

df_cor_unvalid = pd.DataFrame({'succes_mnli_max': [df_unvalid['succes_human'].corr(df_unvalid['succes_mnli_max'])]})
for col in df_unvalid.columns.drop(["succes_human","succes_mnli_max","task_form"]):
    df_cor_unvalid[col]= df_unvalid['succes_human'].corr(df_unvalid[col])
df_cor_unvalid.to_csv("../data/results/unvalid_cor_04_03_22.csv")
df_cor_unvalid

def type_syl(task_form):
    if task_form[0] in ["A", "E"] and task_form[1] in ["A", "E"]:
        return "universal"
    elif task_form[0] in ["O", "I"] and task_form[1] in ["O", "I"]:
        return "relative"
    else:
        return "mixe"

df_valid['type'] = df_valid.task_form.apply(lambda x: type_syl(x))
df_unvalid['type'] = df_unvalid.task_form.apply(lambda x: type_syl(x))

df_valid = df_valid.drop(['task_form'], axis=1)
df_unvalid = df_unvalid.drop(['task_form'], axis=1)

df_valid.groupby('type').mean().to_csv("../data/results/valid_mean_04_03_22.csv")
df_valid.groupby('type').mean()

df_unvalid.groupby('type').mean().to_csv("../data/results/unvalid_mean_04_03_22.csv")
df_unvalid.groupby('type').mean()
