==> notebook/11_similarity.ipynb <==
import pandas as pd
import numpy as np
import ast
from utils import choice_form, implicatur

df = pd.read_csv("../data/intermediate/df_merge_results_with_pred.csv")

df.head()

df.shape

df.task_form.value_counts()

# on retourne 1 si l'humain et la machine ont la meme prédiction, sinon 0
def similarity(human_response, prediction):
    human_response = ast.literal_eval(human_response)
    prediction = ast.literal_eval(prediction)
    return 1 if choice_form[prediction[0]] == human_response[0] else 0


df['similarity_simple'] = df[["human_response","best_prediction"]].apply(lambda x: similarity(x[0],x[1]), axis=1)
df['similarity_mnli'] = df[["human_response","best_few_shot"]].apply(lambda x: similarity(x[0],x[1]), axis=1)

def humanright(human_response):
    human_response =  ast.literal_eval(human_response)
    return human_response[1]

df['human_right'] = df.human_response.apply(lambda x: humanright(x))

df.head()

df_similarity_mnli_true_origin = df[df.human_right==True].groupby(['has_conclusion','task_form']) \
       .agg({'id_seq':'size', 'similarity_mnli':'mean'}) 
df_similarity_mnli_true_origin

df_similarity_mnli_true_origin = df[df.human_right==True].groupby(['has_conclusion','task_form']) \
       .agg({'id_seq':'size', 'similarity_mnli':'mean'}) \
       .rename(columns={'id_seq':'count','similarity_mnli':'mean_mnli_human_right'}) \
       .reset_index()

df_similarity_mnli_false = df[df.human_right==False].groupby('task_form') \
       .agg({'similarity_mnli':'mean'}) \
       .rename(columns={'similarity_mnli':'mean_mnli_human_false'}) \
       .reset_index()

df_similarity_simple_true = df[df.human_right==True].groupby('task_form') \
       .agg({'similarity_simple':'mean'}) \
       .rename(columns={'similarity_simple':'mean_simple_human_right'}) \
       .reset_index()

df_similarity_simple_false = df[df.human_right==False].groupby('task_form') \
       .agg({'similarity_simple':'mean'}) \
       .rename(columns={'similarity_simple':'mean_simple_human_false'}) \
       .reset_index()

df_final = df_similarity_mnli_true_origin.merge(df_similarity_mnli_false, on="task_form")
df_final = df_final.merge(df_similarity_simple_true, on="task_form")
df_final = df_final.merge(df_similarity_simple_false, on="task_form")


df_valid = pd.read_csv("../data/results/result_valid_30_01_22.csv")
df_final_valid = df_final[df_final["has_conclusion"]==True].merge(df_valid, on="task_form")
df_final_valid = df_final_valid[[ "task_form","count", "mean_mnli_human_right","mean_mnli_human_false", "mean_simple_human_right", "mean_simple_human_false", "succes_human" ]]
df_final_valid = df_final_valid.merge(df_valid[["task_form","succes_mnli_00","succes_simple_0.9999965" ]],on="task_form")
df_final_valid['implicatur'] = df_final_valid.task_form.apply(lambda x : True if x in implicatur else False)

df_final_valid.head()

df_final_valid.to_csv("../data/results/result_similarity_valid.csv")

df_unvalid = pd.read_csv("../data/results/result_unvalid_30_01_22.csv")
df_final_unvalid = df_final[df_final["has_conclusion"]==False].merge(df_unvalid, on="task_form")
df_final_unvalid = df_final_unvalid[[ "task_form","count", "mean_mnli_human_right","mean_mnli_human_false", "mean_simple_human_right", "mean_simple_human_false", "succes_human" ]]

df_final_unvalid = df_final_unvalid.merge(df_unvalid[["task_form","succes_mnli_00","succes_simple_0.9999965" ]],on="task_form")
df_final_unvalid['implicatur'] = df_final_unvalid.task_form.apply(lambda x : True if x in implicatur else False)

df_final_unvalid.head()

df_final_unvalid.to_csv("../data/results/result_similarity_unvalid.csv")
