==> notebook/12_a_few_shots.ipynb <==
# I Create the table to train

import pandas as pd
from utils import choice_form

df_train = pd.read_csv("../data/intermediate/df_to_train.csv")
df_conclusion  = pd.read_csv("../data/intermediate/df_conclusion.csv")

df_train = df_train.merge(df_conclusion, on = "id_seq" )

df_train.head()

df = df_train[df_train["has_conclusion"]==True]

df = df[["id_seq","sentenced","choice_str"]]

df.to_csv("../data/intermediate/df_valid_to_train_fs.csv")

# II Analyse des résultats

df_trained = pd.read_csv("../data/intermediate/few_shot_trained_full.csv")
df_conclusion  = pd.read_csv("../data/intermediate/df_conclusion.csv")

df_trained = df_trained.merge(df_conclusion, on = "id_seq" )

df_trained.head()

import ast

def select_best(prediction,choices):
    """ determine the answer of the model as the conclusion with the 
    highest probability
    """
    # We clean the prediction
    prediction = prediction[1:]
    if isinstance(choices,str):
        choices = ast.literal_eval(choices)
    
    for conclusion in choices:
        if prediction == conclusion:
            return choices.index(prediction)
        
    return - 1  


# We determine what is the form of the few shot answer
df_trained['index_choice'] = df_trained[["few_shot_pred","choice_str"]].apply(lambda x: select_best(x[0],x[1]), axis=1)
df_trained['form_choice'] = df_trained.index_choice.apply(lambda x: choice_form[x])

import ast
df_trained['few_shot_succes'] = df_trained[["index_choice","index_conclusion"]].apply(lambda x: 1 if x[0] in ast.literal_eval(x[1]) else 0, axis=1)

df_result_fewshot = df_trained[['task_form', 'few_shot_succes']].groupby(['task_form']).mean()
df_result_fewshot['task_form'] = df_result_fewshot.index
df_result_fewshot.to_csv("../data/results/df_fewshot_results.csv" ,index=False)

df = df_trained[["task_form", "form_choice"]]
#df_groupby = df.groupby('task_form').form_choice.value_counts()
df_groupby = df.groupby(['task_form',"form_choice"]).size().reset_index(name='counts')


df_groupby

df_groupby.to_csv("../data/results/df_fewshot_group_by.csv" ,index=False)
