==> notebook/13_merge_answer.ipynb <==
import pandas as pd

df_valid = pd.read_csv("../data/results/result_valid_30_01_22.csv")
df_unvalid = pd.read_csv("../data/results/result_unvalid_30_01_22.csv")


df = pd.concat([df_valid,df_unvalid])
df = df.drop(columns=["Unnamed: 0"])

from utils import implicatur
df_conclusion = pd.read_csv("../data/results/conclusion.csv")
df_conclusion["has_conclusion"]= df_conclusion.conclusion.apply(lambda x : False if x == "[NVC]" else True )
df_conclusion['implicatur'] = df_conclusion.task_form.apply(lambda x : True if x in implicatur else False)
df_conclusion.drop(columns=["conclusion"])

df = pd.merge(df, df_conclusion, on=["task_form"])
#df.to_csv("../data/app/df_results.csv")

df.columns
